26/11/2023

+++ database/migrations/2023_08_06_231807_create_houses_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
         Schema::create('houses', function (Blueprint $table) {
             $table->id();
             $table->string('house_name');
-            $table->integer('house_price');
+            $table->bigInteger('house_price');
             $table->text('house_description');
             $table->integer('house_room_count');
             $table->integer('house_garage_count');
